+++ src/generator/with_rand09.rs
@@ -3,7 +3,7 @@
 #![cfg(feature = "rand09")]
 #![deprecated(since = "3.6.0", note = "use a newer version of `rand` crate")]
 
-use super::{RandSource, Scru128Generator, StdSystemTime};
+use super::{Generator, RandSource, StdSystemTime};
 use rand_core09::RngCore;
 
 /// An adapter that implements [`RandSource`] for [`RngCore`] types.
@@ -16,7 +16,7 @@ impl<T: RngCore> RandSource for Adapter<T> {
     }
 }
 
-impl<T: RngCore> Scru128Generator<Adapter<T>> {
+impl<T: RngCore> Generator<Adapter<T>> {
     /// Creates a generator object with a specified random number generator that implements
     /// [`RngCore`] from `rand` (v0.9) crate. The specified random number generator should be
     /// cryptographically strong and securely seeded.
@@ -27,9 +27,7 @@ impl<T: RngCore> Scru128Generator<Adapter<T>> {
     /// # #[cfg(feature = "default_rng")]
     /// # {
     /// # use rand09 as rand;
-    /// use scru128::Scru128Generator;
-    ///
-    /// let mut g = Scru128Generator::with_rand09(rand::rng());
+    /// let mut g = scru128::Generator::with_rand09(rand::rng());
     /// println!("{}", g.generate());
     /// # }
     /// ```