+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## v3.6.0 - unreleased
 
+- Renamed `Scru128Generator` and `Scru128Id` to `Generator` and `Id`,
+  respectively, so they can be used as `scru128::Generator` and `scru128::Id`.
+  Old names are left as deprecated aliases for backward compatibility.
 - Deprecated `default_rng` cargo feature and related APIs, most notably, the
   `new()` constructor and default type parameter `R` of `Scru128Generator`.
 - Deprecated `rand09` cargo feature and the integration with `rand` crate v0.9.
@@ -10,7 +13,7 @@
 - Added `Scru128Generator::iter()` to supersede the direct `Iterator`
   implementation for `Scru128Generator`.
 - Tweaked the `Debug` representation of `Scru128Generator`.
-- Removed deprecated `Scru128Rng` trait alias from documentation.
+- Hid deprecated `Scru128Rng` trait alias from documentation.
 - Upgraded the dev dependency on `rand` from v0.9 to v0.10.
 - Minor refactoring.